# daemon/docs/swagger.yaml
basePath: /
definitions:
...
  waypaper-engine_daemon_internal_handler_httpjson.APIError:
    properties:
      code:
        type: integer
      details:
        type: string
      error:
        type: string
      error_code:
        type: string
      meta:
        additionalProperties: {}
        type: object
    type: object
...
  waypaper-engine_daemon_internal_monitor.Monitor:
    properties:
      height:
        description: Height is the vertical resolution in pixels.
        type: integer
      name:
        description: Name is the output identifier as reported by the compositor (e.g.
          "HDMI-A-1", "eDP-1").
        type: string
      refresh_rate:
        description: RefreshRate is the monitor's refresh rate in Hz (e.g. 60.0, 144.0).
        type: number
      scale:
        description: Scale is the output scale factor (e.g. 1.0, 1.5, 2.0).
        type: number
      transform:
        description: |-
          Transform is the rotation/reflection applied to the output.
          0=normal, 1=90°, 2=180°, 3=270°, 4=flipped, 5=flipped-90°, 6=flipped-180°, 7=flipped-270°.
        type: integer
      width:
        description: Width is the horizontal resolution in pixels.
        type: integer
      x:
        description: |-
          X is the horizontal position in the compositor's coordinate space.
          Used for extend mode to compute which slice of the image maps to this monitor.
        type: integer
      "y":
        description: Y is the vertical position in the compositor's coordinate space.
        type: integer
    type: object
...
info:
  contact: {}
  description: HTTP API served over a Unix socket for the waypaper-engine wallpaper
    manager.
  title: waypaper-engine daemon
  version: 3.0.0
paths:
...
  /backends/{name}/activate:
    post:
      parameters:
      - description: Backend name
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Activate a backend
      tags:
      - backends
...
  /config:
    get:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Get full config
      tags:
      - config
    patch:
      parameters:
      - description: Section patches
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            additionalProperties: true
            type: object
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Patch config sections
      tags:
      - config
  /config/{section}:
    get:
      parameters:
      - description: Section name
        in: path
        name: section
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Get a config section
      tags:
      - config
    patch:
      parameters:
      - description: Section name
        in: path
        name: section
        required: true
        type: string
      - description: Section values
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Patch a config section
      tags:
      - config
  /config/backends/{backend}:
    get:
      parameters:
      - description: Backend name
        in: path
        name: backend
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Get backend config
      tags:
      - config
    patch:
      parameters:
      - description: Backend name
        in: path
        name: backend
        required: true
        type: string
      - description: Backend config patch
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Patch backend config
      tags:
      - config
...
  /monitors:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/waypaper-engine_daemon_internal_monitor.Monitor'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: List connected monitors
      tags:
      - monitors
  /monitors/{name}:
    get:
      parameters:
      - description: Monitor name
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_monitor.Monitor'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/waypaper-engine_daemon_internal_handler_httpjson.APIError'
      summary: Get a monitor by name
      tags:
      - monitors
...
swagger: "2.0"